requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
docxtpl==0.10.0
idna==2.9
importlib-metadata==1.6.1
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
Pillow==6.2.2
pkg-resources==0.0.0
py==1.4.22
pymongo==3.10.1
pytest==2.6.0
python-decouple==3.3
python-docx==0.8.10
requests==2.23.0
six==1.15.0
tqdm==4.46.1
ui==0.1.4
urllib3==1.25.9
wget==3.2
zipp==3.1.0
